refactor(logstreams): move interface to logstream
Moves the `ImmutableRecordBatchEntry` to the `logstreams` module, and
renames it to `ImmutableAppendEntry`. This is a counterpart of
`LoggedEvent` meant strictly for appending.

While this is still exposed in the public API of `stream-platform`, it
will be eventually removed and should be used only by the stream
platform implementation (which it already is).
