+++ stream-platform/src/main/java/io/camunda/zeebe/stream/impl/records/RecordBatch.java
@@ -7,13 +7,13 @@
  */
 package io.camunda.zeebe.stream.impl.records;
 
+import io.camunda.zeebe.logstreams.log.LogAppendEntry;
 import io.camunda.zeebe.protocol.record.RecordType;
 import io.camunda.zeebe.protocol.record.RejectionType;
 import io.camunda.zeebe.protocol.record.ValueType;
 import io.camunda.zeebe.protocol.record.intent.Intent;
 import io.camunda.zeebe.stream.api.records.ExceededBatchRecordSizeException;
 import io.camunda.zeebe.stream.api.records.ImmutableRecordBatch;
-import io.camunda.zeebe.stream.api.records.ImmutableRecordBatchEntry;
 import io.camunda.zeebe.stream.api.records.MutableRecordBatch;
 import io.camunda.zeebe.stream.api.records.RecordBatchSizePredicate;
 import io.camunda.zeebe.util.Either;
@@ -26,7 +26,7 @@ import java.util.function.Consumer;
 
 public final class RecordBatch implements MutableRecordBatch {
 
-  final List<ImmutableRecordBatchEntry> recordBatchEntries = new ArrayList<>();
+  final List<LogAppendEntry> recordBatchEntries = new ArrayList<>();
   private int batchSize;
   private final RecordBatchSizePredicate recordBatchSizePredicate;
 
@@ -81,17 +81,17 @@ public final class RecordBatch implements MutableRecordBatch {
   }
 
   @Override
-  public Iterator<ImmutableRecordBatchEntry> iterator() {
+  public Iterator<LogAppendEntry> iterator() {
     return recordBatchEntries.iterator();
   }
 
   @Override
-  public void forEach(final Consumer<? super ImmutableRecordBatchEntry> action) {
+  public void forEach(final Consumer<? super LogAppendEntry> action) {
     recordBatchEntries.forEach(action);
   }
 
   @Override
-  public Spliterator<ImmutableRecordBatchEntry> spliterator() {
+  public Spliterator<LogAppendEntry> spliterator() {
     return recordBatchEntries.spliterator();
   }
 }
